Fix bug with build
