+++ mk/mk.deps
@@ -796,7 +796,29 @@ src/rssparser.o: src/rssparser.cpp include/rssparser.h \
 src/ruststring.o: src/ruststring.cpp include/ruststring.h
 src/scopemeasure.o: src/scopemeasure.cpp include/scopemeasure.h \
  target/cxxbridge/libnewsboat-ffi/src/scopemeasure.rs.h
-src/searchresultsformaction.o: src/searchresultsformaction.cpp
+src/searchresultsformaction.o: src/searchresultsformaction.cpp \
+ include/searchresultsformaction.h include/configcontainer.h \
+ include/configactionhandler.h include/itemlistformaction.h \
+ 3rd-party/optional.hpp include/history.h \
+ target/cxxbridge/libnewsboat-ffi/src/history.rs.h \
+ include/listformaction.h include/formaction.h include/keymap.h \
+ include/stflpp.h include/listwidget.h include/listformatter.h \
+ include/regexmanager.h include/matcher.h filter/FilterParser.h \
+ include/regexowner.h include/view.h include/colormanager.h \
+ include/controller.h include/cache.h include/configparser.h \
+ include/feedcontainer.h include/filtercontainer.h include/fslock.h \
+ target/cxxbridge/libnewsboat-ffi/src/fslock.rs.h include/opml.h \
+ include/fileurlreader.h include/urlreader.h include/utils.h \
+ 3rd-party/expected.hpp include/logger.h config.h include/strprintf.h \
+ target/cxxbridge/libnewsboat-ffi/src/logger.rs.h \
+ target/cxxbridge/libnewsboat-ffi/src/utils.rs.h include/queuemanager.h \
+ include/reloader.h include/remoteapi.h include/rssignores.h \
+ include/rssitem.h include/matchable.h include/dirbrowserformaction.h \
+ include/filesystembrowser.h include/feedlistformaction.h \
+ include/filebrowserformaction.h include/htmlrenderer.h \
+ include/textformatter.h include/statusline.h include/rssfeed.h \
+ include/keymap.h include/view.h include/configcontainer.h \
+ include/itemlistformaction.h include/regexmanager.h include/rssfeed.h
 src/selectformaction.o: src/selectformaction.cpp \
  include/selectformaction.h include/filtercontainer.h \
  include/configactionhandler.h include/formaction.h include/history.h \
